Move pyenv utility functions to their own module
