+++ spyder/utils/programs.py
@@ -31,7 +31,7 @@ import psutil
 
 # Local imports
 from spyder.config.base import (is_stable_version, running_under_pytest,
-                                get_home_dir, running_in_mac_app)
+                                get_home_dir)
 from spyder.config.utils import is_anaconda
 from spyder.py3compat import PY2, is_text_string, to_text_string
 from spyder.utils import encoding
@@ -39,7 +39,6 @@ from spyder.utils.misc import get_python_executable
 
 
 HERE = osp.abspath(osp.dirname(__file__))
-WINDOWS = os.name == 'nt'
 
 
 class ProgramError(Exception):
@@ -105,7 +104,7 @@ def is_program_installed(basename):
 
         req_paths.extend(pyenv + conda)
 
-    elif WINDOWS:
+    elif os.name == 'nt':
         pyenv = [osp.join(home, '.pyenv', 'pyenv-win', 'bin')]
 
         a = [home, 'C:\\', osp.join('C:\\', 'ProgramData')]
@@ -1055,48 +1054,6 @@ def is_spyder_process(pid):
         return any(conditions)
     except (psutil.NoSuchProcess, psutil.AccessDenied):
         return False
-
-
-def get_pyenv_path(name):
-    """Return the complete path of the pyenv."""
-    home = get_home_dir()
-    if WINDOWS:
-        path = osp.join(
-            home, '.pyenv', 'pyenv-win', 'versions', name, 'python.exe')
-    elif name == '':
-        path = osp.join(home, '.pyenv', 'shims', 'python')
-    else:
-        path = osp.join(home, '.pyenv', 'versions', name, 'bin', 'python')
-    return path
-
-
-def get_list_pyenv_envs():
-    """Return the list of all pyenv envs found in the system."""
-    env_list = {}
-    pyenv = find_program('pyenv')
-    if pyenv is None:
-        return env_list
-
-    cmdstr = ' '.join([pyenv, 'versions', '--bare', '--skip-aliases'])
-    try:
-        out, err = run_shell_command(cmdstr, env={}).communicate()
-        out = out.decode()
-        err = err.decode()
-    except Exception:
-        out = ''
-        err = ''
-    out = out.split('\n')
-    for env in out:
-        data = env.split(osp.sep)
-        path = get_pyenv_path(data[-1])
-        if data[-1] == '':
-            name = 'internal' if running_in_mac_app(path) else 'system'
-        else:
-            name = 'pyenv: {}'.format(data[-1])
-        version = (
-            'Python 2.7' if data[-1] == '' else 'Python {}'.format(data[0]))
-        env_list[name] = (path, version)
-    return env_list
 
 
 def get_interpreter_info(path):
